requirements.txt
ahk==1.8.4
ahk-binary==2023.9.0
customtkinter==5.2.2
darkdetect==0.8.0
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.5.1
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==25.1.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
pytweening==1.2.0
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
sympy==1.14.0
tifffile==2025.6.11
tk==0.1.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
